pyproject.toml
[project]
name = "igc-visualizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "rerun-sdk>=0.23.3",
]
